% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_activities_list.R
\name{clean_activities_list}
\alias{clean_activities_list}
\title{clean_activities_list}
\usage{
clean_activities_list(activities = NULL)
}
\arguments{
\item{activities}{json activities returnd by get_activities}
}
\value{
activities of the page
}
\description{
a function to clean strava actities list returned from strava API
}
\examples{
\dontrun{
clean_activities_list(my_activities)
}
}
