% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-activities.R
\name{cleanActivities}
\alias{cleanActivities}
\title{cleanActivities}
\usage{
cleanActivities(activities, myGearList)
}
\arguments{
\item{activities}{json activities returnd by getActivities}

\item{myGearList}{}
}
\description{
a function to clean strava actities
}
