% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-activities-list.R
\name{cleanActivitiesList}
\alias{cleanActivitiesList}
\title{cleanActivitiesList}
\usage{
cleanActivitiesList(activities = NULL)
}
\arguments{
\item{activities}{json activities returnd by getActivities}
}
\value{
activities of the page
}
\description{
a function to clean strava actities
}
\examples{
\dontrun{
cleanActivities(my_activities)
}
}
