% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decodePolyline.R
\name{decodePolyline}
\alias{decodePolyline}
\title{Decode Google polyline into latitude and longitude}
\usage{
decodePolyline(polyline = NULL)
}
\arguments{
\item{polyline}{character element of Google polyline of an activity}
}
\value{
A vector of latitudes an longitudes in decimals separated by a comma
}
\description{
This function converts the polyline into latitude and longitude coordinates.
}
\examples{
\dontrun{
decode_Polyline()}
}
\seealso{
\url{https://s4rdd.blogspot.co.uk/2012/12/google-maps-api-decoding-polylines-for.html?showComment=1473004506791#c3610119369153401460}
}
\author{
Daniel Padfield
}
\concept{notoken}
