% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_get_athelete_stats.R
\name{def_get_athlete_stats}
\alias{def_get_athlete_stats}
\title{def_get_athlete_stats}
\usage{
def_get_athlete_stats(athlete_id = NULL)
}
\arguments{
\item{athlete_id}{the strava id of the athelete (must be the one logged)}
}
\description{
def_get_athlete_stats
}
