% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_list_activities.R
\name{def_list_activities}
\alias{def_list_activities}
\title{def_list_activities}
\usage{
def_list_activities(
  activities_per_page = 200,
  page_number = 1,
  after_date = NULL,
  before_date = NULL
)
}
\arguments{
\item{activities_per_page}{max number of item to return (must be 200 max)}

\item{page_number}{the specific page number to extract}

\item{after_date}{return activities after date as YYYY-MM-DD}

\item{before_date}{return activities before date as YYYY-MM-DD}
}
\description{
def_list_activities
}
