% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_explore_segments.R
\name{explore_segments}
\alias{explore_segments}
\title{explore_segments}
\usage{
explore_segments(
  mbr_object = NULL,
  southwest_lat = NULL,
  southwest_long = NULL,
  northeast_lat = NULL,
  northeast_long = NULL,
  activity_type = NULL,
  min_climbing_cat = NULL,
  max_climbing_cat = NULL
)
}
\arguments{
\item{mbr_object}{a mountainbiker object}

\item{southwest_lat}{southwest corner latitutde}

\item{southwest_long}{southwest corner longitude}

\item{northeast_lat}{northeast corner latitude}

\item{northeast_long}{northeast corner longitude}

\item{activity_type}{Desired activity type. May take one of the
following values: running, riding}

\item{min_climbing_cat}{int the minimum climbing category}

\item{max_climbing_cat}{int the maximum climbing category}
}
\value{
my_segments as list with activity details
}
\description{
Returns the given activity that is owned by the authenticated athlete.
https://developers.strava.com/playground/#/Activities/getActivityById
}
\examples{
\dontrun{
# my_mbr is a mountainbikeR object

my_mbr$explore_segments(
     activity_id = 12345,
     include_all_efforts = TRUE
)
}
}
