% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPath.R
\name{extractPath}
\alias{extractPath}
\title{extractPath}
\usage{
extractPath(polyline = NULL, id = NULL, date_time = NULL)
}
\arguments{
\item{polyline}{a polyline as per google algo}

\item{id}{the associated id of the polyline (the segment or strava activity id)}

\item{date_time}{the YYYY-MM-DD date}
}
\value{
a data.frame with lon, lat, point_id, id, date as column
}
\description{
a function to extract a path from a polyline coded GPS track which returns
a data.table
}
\examples{
\dontrun{
extractPath(
      polyline = myPolyline,
      id = 123456
      date_time = "2018-01-12"
)
}
}
\seealso{
\url{https://developers.google.com/maps/documentation/utilities/polylinealgorithm}
}
