% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-segment-efforts.R
\name{extractSegmentEfforts}
\alias{extractSegmentEfforts}
\title{extractSegmentEfforts}
\usage{
extractSegmentEfforts(activity.details)
}
\arguments{
\item{activity.details}{}
}
\description{
return all segment_efforts summary data.frame from an activity detailled list
}
