% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_get_activities.R
\name{get_activities}
\alias{get_activities}
\title{get_activities}
\source{
\url{https://developers.strava.com/docs/reference/#api-Activities-getLoggedInAthleteActivities}
}
\usage{
get_activities(
  mbr_object = NULL,
  activities_per_page = 200,
  page_number = 1,
  after_date = NULL,
  before_date = NULL
)
}
\arguments{
\item{mbr_object}{a mountainbikeR object}

\item{activities_per_page}{max number of item to return (must be 200 max)}

\item{page_number}{the specific page number to extract}

\item{after_date}{return activities after date as YYYY-MM-DD}

\item{before_date}{return activities before date as YYYY-MM-DD}
}
\value{
the activities database as json
}
\description{
List all activities from strava api. Returns the activities of an athlete
for a specific identifier. Requires activity:read.
Only Me activities will be filtered out unless requested by a token
with activity:read_all.
}
\examples{
\dontrun{
# my_mbr is a mountainbikeR objecti previously created
my_mbr$get_activities(
  activities_per_page = 200
  , page_number = 2
 , after_date = "2018-02-20"
 , before_date = "2018-03-01"
  )
}
}
