% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_get_activity_zones.R
\name{get_activity_zones}
\alias{get_activity_zones}
\title{get_activity}
\usage{
get_activity_zones(mbr_object = NULL, activity_id = NULL)
}
\arguments{
\item{mbr_object}{a mountainbiker object}

\item{activity_id}{the strava id of the activity}
}
\value{
my_activity_zones as list with activity details
}
\description{
Returns the zones of a given activity.
}
\examples{
\dontrun{
# my_mbr as mountainbikeR object

my_mbr$get_activty_zone(
     activity_id = 12345
)
}
}
\seealso{
\url{https://developers.strava.com/docs/reference/#api-Activities-getZonesByActivityId}
}
