% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_get_athlete.R
\name{get_athlete}
\alias{get_athlete}
\title{get_athlete}
\usage{
get_athlete(mbr_object = NULL)
}
\arguments{
\item{mbr_object}{a mountainbiker object}
}
\value{
a list with athlete details
}
\description{
Returns the logged athlete
}
\examples{
\dontrun{
# my_mbr is an existing mountainbikeR object

my_mbr$get_athlete(
)
}
}
\seealso{
\url{https://developers.strava.com/docs/reference/#api-Athletes-getLoggedInAthlete}
}
