% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_get_athelete_stats.R
\name{get_athlete_stats}
\alias{get_athlete_stats}
\title{get_athlete_stats}
\usage{
get_athlete_stats(mbr_object = NULL, athlete_id = NULL)
}
\arguments{
\item{mbr_object}{a mountainbiker object}

\item{athlete_id}{the strava id of the athelete (must be the one logged)}
}
\value{
as list with athelete stats
}
\description{
Returns the activity stats of an athlete. Only includes data from activities
 set to Everyone visibilty.
}
\examples{
\dontrun{
# my_mbr is a mountainbikeR object defined beforehead

my_mbr$get_athlete_stats(
     athelete_id = 12345
)
}
}
\seealso{
\url{https://developers.strava.com/docs/reference/#api-Athletes-getStats}
}
