% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_get_athelete_zones.R
\name{get_athlete_zones}
\alias{get_athlete_zones}
\title{get_athlete_zones}
\usage{
get_athlete_zones(mbr_object = NULL)
}
\arguments{
\item{mbr_object}{a mountainbiker object}
}
\value{
as list with athelete zones
}
\description{
Returns the activity zones of an athlete. Only includes data from activities
 set to Everyone visibilty.
}
\examples{
\dontrun{
# my_mbr is a mountainbikeR object defined beforehead

my_mbr$get_athlete_zones()
}
}
\seealso{
\url{https://developers.strava.com/docs/reference/#api-Athletes-getLoggedInAthleteZones}
}
