% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_get_gear.R
\name{get_gear}
\alias{get_gear}
\title{getGear}
\usage{
get_gear(mbr_object, gear_id)
}
\arguments{
\item{mbr_object}{a mountainbiker object}

\item{gear_id}{the id of the gear like b34565 for a bike}
}
\value{
the gear data as json
}
\description{
Get all gear detail from strava api from gears id
}
\examples{
\dontrun{
get_gear(
  mbr_object = myObject
  , gear_id = myGearId
  )
}
}
