% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_get_segment.R
\name{get_segment}
\alias{get_segment}
\title{get_segment}
\usage{
get_segment(mbr_object = NULL, segment_id = NULL)
}
\arguments{
\item{mbr_object}{a mountainbiker object}

\item{segment_id}{the strava id of the segment}
}
\value{
my_segment as list with segment details
}
\description{
Returns the specified segment. read_all scope required in order to
retrieve athlete-specific segment information, or to retrieve private
segments.
}
\examples{
\dontrun{
# my_mbr is a mountainbikeR object

my_mbr$get_segment(
     segment_id = 12345
)
}
}
\seealso{
\url{https://developers.strava.com/docs/reference/#api-Segments-getSegmentById}
}
