% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_get_segment_effort.R
\name{get_segment_effort}
\alias{get_segment_effort}
\title{get_segment_effort}
\usage{
get_segment_effort(mbr_object = NULL, effort_id = NULL)
}
\arguments{
\item{effort_id}{the id of the specific effort to collect}
}
\value{
my_effort containing effort details
}
\description{
Returns a segment effort from an activity that is owned by the
authenticated athlete.
}
\examples{
\dontrun{
# my_mbr is a mountainbikeR object

my_mbr$get_segment_effort(
     effort_id = 12345
)
}
}
\seealso{
\url{https://developers.strava.com/docs/reference/#api-SegmentEfforts-getSegmentEffortById}

\url{https://developers.strava.com/playground/#/SegmentEfforts/getSegmentEffortById}
}
