% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_get_segment_effort_streams.R
\name{get_segment_effort_streams}
\alias{get_segment_effort_streams}
\title{get_segment_effort_streams}
\usage{
get_segment_effort_streams(
  mbr_object = NULL,
  effort_id = NULL,
  streams_types = NULL
)
}
\arguments{
\item{effort_id}{The identifier of the segment effort.}

\item{streams_types}{vector of esired stream types. May take the
following values:
"time", "latlng", "distance", "altitude", "velocity_smooth",
"heartrate", "cadence", "watts", "temp", "moving", "grade_smooth"}
}
\value{
my_segment_effort_streams as list with streams details
}
\description{
Returns a set of streams for a segment effort completed by the
authenticated athlete. Requires read_all scope.
}
\examples{
\dontrun{
# my_mbr is a mountainbikeR object

my_mbr$get_segment_effort_streams(
     effort_id = 12345
     , streams_types = c("heartrate","distance")
)
}
}
\seealso{
\url{https://developers.strava.com/docs/reference/#api-Streams-getSegmentEffortStreams}
}
