% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_get_segment_leaderboard.R
\name{get_segment_leaderboard}
\alias{get_segment_leaderboard}
\title{get_segment_leaderboard}
\usage{
get_segment_leaderboard(
  mbr_object = NULL,
  segment_id = NULL,
  gender = NULL,
  age_group = NULL,
  weight_class = NULL,
  following = NULL,
  club_id = NULL,
  date_range = NULL,
  context_entries = NULL,
  page = NULL,
  per_page = NULL
)
}
\arguments{
\item{segment_id}{int The identifier of the segment leaderboard.}

\item{gender}{string, Filter by gender. May take one of the following
values: M, F}

\item{age_group}{string summit Feature. Filter by age group. May
take one of the following values: 0_19, 20_24, 25_34, 35_44, 45_54,
55_64, 65_69, 70_74, 75_plus}

\item{weight_class}{string, filter by weight class. May take one of
the following values: 0_124, 125_149, 150_164, 165_179, 180_199,
200_224, 225_249, 250_plus, 0_54, 55_64, 65_74, 75_84, 85_94, 95_104,
105_114, 115_plus}

\item{following}{Boolean Filter by friends of the authenticated
athlete.}

\item{club_id}{Long,Filter by club id}

\item{date_range}{string, Filter by date range, will be in the
athlete's timezone May take one of the following values:
this_year, this_month, this_week, today}

\item{context_entries}{Integer, ??? not documented}

\item{page}{Integer, Page number}

\item{per_page}{integer, Number of items per page. Defaults to 30.}
}
\value{
my_segment_leaderboard as list with leaderboard
}
\description{
Returns the specified segment leaderboard.
}
\examples{
\dontrun{
# my_mbr is a mountainbikeR object

my_mbr$get_segment_leaderboard(
)
}
}
\seealso{
\url{https://developers.strava.com/docs/reference/#api-Segments-getLeaderboardBySegmentId}
}
