% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_get_segment_streams.R
\name{get_segment_streams}
\alias{get_segment_streams}
\title{get_segment_streams}
\usage{
get_segment_streams(mbr_object = NULL, segment_id = NULL, streams_types = NULL)
}
\arguments{
\item{mbr_object}{a mountainbiker object}

\item{segment_id}{the strava id of the segment}

\item{streams_types}{vector of esired stream types. May take the
following values:
"latlng", "distance", "altitude"}
}
\value{
my_streams as list with streams details
}
\description{
Returns the specified segment streams but only altitude, latlong and distance
read_all scope required in order to retrieve athlete-specific segment
information, or to retrieve private segments.
}
\examples{
\dontrun{
# my_mbr is a mountainbikeR object

my_mbr$get_segment_streams(
     segment_id = 12345
     , streams_types = c("heartrate","distance")
)
}
}
\seealso{
\url{https://developers.strava.com/docs/reference/#api-Segments-getSegmentById}
}
