% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_get_starred_segments.R
\name{get_starred_segments}
\alias{get_starred_segments}
\title{get_starred_segments}
\usage{
get_starred_segments(mbr_object = NULL, page = NULL, per_page = NULL)
}
\arguments{
\item{page}{Integer, Page number}

\item{per_page}{integer, Number of items per page. Defaults to 30.}
}
\value{
my_starred_segments as list with leaderboard
}
\description{
List of the authenticated athlete's starred segments. Private segments are
filtered out unless requested by a token with read_all scope.
}
\examples{
\dontrun{
# my_mbr is a mountainbikeR object

my_mbr$get_starred_segments(
)
}
}
\seealso{
\url{https://developers.strava.com/docs/reference/#api-Segments-getLoggedInAthleteStarredSegments}
}
