% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-activities.R
\name{getActivities}
\alias{getActivities}
\title{getActivities}
\usage{
getActivities(token = NULL)
}
\arguments{
\item{token}{the strava identification token}
}
\value{
the activities database as json
}
\description{
Get all activities from strava api
}
