% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getActivity.R
\name{getActivity}
\alias{getActivity}
\title{getActivity}
\usage{
getActivity(token = NULL, activity_id = NULL,
  include_all_efforts = FALSE)
}
\arguments{
\item{token}{the strava token}

\item{activity_id}{the strava id of the activity}

\item{include_all_efforts}{boolean}
}
\value{
myActivity as list with activity details
}
\description{
Returns the given activity that is owned by the authenticated athlete.
https://developers.strava.com/playground/#/Activities/getActivityById
}
\examples{
\dontrun{
getActivity(
     token = mytoken,
     activity_id = 12345,
     include_all_efforts = TRUE
)
}
}
