% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-activity-details.R
\name{getActivityDetails}
\alias{getActivityDetails}
\title{getActivity}
\usage{
getActivityDetails(activity.id)
}
\arguments{
\item{activity.id}{}
}
\value{
myActivity as list with activity details
}
\description{
Returns the given activity that is owned by the authenticated athlete.
https://developers.strava.com/playground/#/Activities/getActivityById
}
