% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGear.R
\name{getGear}
\alias{getGear}
\title{getGear}
\usage{
getGear(token = NULL, gear.id = NULL)
}
\arguments{
\item{token}{the strava identification token}

\item{gear.id}{the id of the gear}
}
\value{
the gear data as json
}
\description{
Get all gear detail from strava api from gears id
}
\examples{
\dontrun{
getGear(token = myToken, gear.id = myGearId)
}
}
