% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-gear-name.R
\name{getGearName}
\alias{getGearName}
\title{getGearName}
\usage{
getGearName(x, gearList)
}
\arguments{
\item{gearList}{}
}
\value{
the name of the gear for a given id
}
\description{
using a gear list database
return gear name from a list of gear id
}
