% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-stream.R
\name{getStream}
\alias{getStream}
\title{getStream}
\usage{
getStream(path2stream, hold.time = 1.8)
}
\arguments{
\item{path2stream}{API path call}

\item{hold.time}{holding time between retry}
}
\description{
get steam from one specific activity
}
