% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-streams.R
\name{getStreams}
\alias{getStreams}
\title{getStreams}
\source{
\url{https://developers.strava.com/docs/reference/#api-Streams-getActivityStreams}
}
\usage{
getStreams(token = NULL, activity_id = NULL, stream_keys = NULL)
}
\arguments{
\item{token}{the strava authorization token}

\item{activity_id}{integer, The identifier of the activity}

\item{stream_types}{a vector of Desired stream types, see stream_types}
}
\description{
get streams from one specific activity
}
\examples{

}
