% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-stream-path.R
\name{makeStreamPath}
\alias{makeStreamPath}
\title{makeStreamPath}
\usage{
makeStreamPath(activity.id = NULL, stream.type = NULL)
}
\arguments{
\item{activity.id}{the id of the activity}

\item{stream.type}{the type of the stream}
}
\value{
the full API path
}
\description{
build the API url path
}
