% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6class.R
\name{mountainbiker_initializer}
\alias{mountainbiker_initializer}
\title{R6 Class for a strava application with authorization}
\description{
This strava class is designed to interract with strava API using
a set of methods.
A mountainbikeR object must be initialized using the new method then
all other method will be availabe
}
\details{
TBD
}
\examples{

## ------------------------------------------------
## Method `mountainbiker_initializer$new`
## ------------------------------------------------

\dontrun{
my_mb <- mountainbiker_initializer$new(
    my_app_name = "mycool_app"
    , my_client_id = 77777
    , my_client_secret = "a super secret string"
    , my_app_scope = "read,read_all,activity:read_all" #"public"
    , cache = FALSE
    )
}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{token}}{Shows generated token.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{mountainbiker_initializer$new()}}
\item \href{#method-get_activities}{\code{mountainbiker_initializer$get_activities()}}
\item \href{#method-get_activity}{\code{mountainbiker_initializer$get_activity()}}
\item \href{#method-get_athlete}{\code{mountainbiker_initializer$get_athlete()}}
\item \href{#method-get_athlete_stats}{\code{mountainbiker_initializer$get_athlete_stats()}}
\item \href{#method-get_athlete_zones}{\code{mountainbiker_initializer$get_athlete_zones()}}
\item \href{#method-get_gear}{\code{mountainbiker_initializer$get_gear()}}
\item \href{#method-clone}{\code{mountainbiker_initializer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new moutainbiker object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mountainbiker_initializer$new(
  my_app_name = NULL,
  my_client_id = NULL,
  my_client_secret = NULL,
  my_app_scope = "read,read_all,activity:read_all",
  cache = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{my_app_name}}{chr string, the name of the app defined in strava}

\item{\code{my_client_id}}{chr string, the ID of the app like 27881}

\item{\code{my_client_secret}}{chr string, the secret as defined in strava}

\item{\code{my_app_scope}}{chr string for requested scopes, as a comma delimited}

\item{\code{cache}}{a logical TRUE to keep cached token, false for not.
string, e.g. "activity:read_all,activity:write"}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `mountainbikeR object` object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
my_mb <- mountainbiker_initializer$new(
    my_app_name = "mycool_app"
    , my_client_id = 77777
    , my_client_secret = "a super secret string"
    , my_app_scope = "read,read_all,activity:read_all" #"public"
    , cache = FALSE
    )
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_activities"></a>}}
\if{latex}{\out{\hypertarget{method-get_activities}{}}}
\subsection{Method \code{get_activities()}}{
List all activities from strava api. Returns the activities of an athlete
for a specific identifier. Requires activity:read.
Only Me activities will be filtered out unless requested by a token
with activity:read_all.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mountainbiker_initializer$get_activities(
  activities_per_page = 200,
  page_number = 1,
  after_date = NULL,
  before_date = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{activities_per_page}}{max number of item to return (must be 200 max)}

\item{\code{page_number}}{the specific page number to extract}

\item{\code{after_date}}{return activities after date as YYYY-MM-DD}

\item{\code{before_date}}{return activities before date as YYYY-MM-DD}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_activity"></a>}}
\if{latex}{\out{\hypertarget{method-get_activity}{}}}
\subsection{Method \code{get_activity()}}{
Returns the given activity that is owned by the authenticated athlete.
https://developers.strava.com/playground/#/Activities/getActivityById
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mountainbiker_initializer$get_activity(
  activity_id = NULL,
  include_all_efforts = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{activity_id}}{the strava id of the activity}

\item{\code{include_all_efforts}}{boolean}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_athlete"></a>}}
\if{latex}{\out{\hypertarget{method-get_athlete}{}}}
\subsection{Method \code{get_athlete()}}{
Get informations about the athlete
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mountainbiker_initializer$get_athlete()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_athlete_stats"></a>}}
\if{latex}{\out{\hypertarget{method-get_athlete_stats}{}}}
\subsection{Method \code{get_athlete_stats()}}{
Get informations about the athlete stats
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mountainbiker_initializer$get_athlete_stats(athlete_id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{athlete_id}}{The identifier of the athlete. Must match the
authenticated athlete.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_athlete_zones"></a>}}
\if{latex}{\out{\hypertarget{method-get_athlete_zones}{}}}
\subsection{Method \code{get_athlete_zones()}}{
Get informations about the athlete heart rate zone
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mountainbiker_initializer$get_athlete_zones()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_gear"></a>}}
\if{latex}{\out{\hypertarget{method-get_gear}{}}}
\subsection{Method \code{get_gear()}}{
Get athlete gear description from gear id
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mountainbiker_initializer$get_gear(gear_id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gear_id}}{the id of the gear to get}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mountainbiker_initializer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
