% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-streams_type.R
\docType{data}
\name{streams_type}
\alias{streams_type}
\title{Strava Streams Types}
\format{A list of available streams types
\describe{
  \item{type}{Type of streams}
  \item{dataype}{Type of data}
  \item{units}{Units of the stream}
}}
\source{
\url{https://developers.strava.com}
}
\usage{
streams_type
}
\description{
A dataset containing the strava stream types and their units
}
\keyword{datasets}
