% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_workout_types.R
\docType{data}
\name{workout_types}
\alias{workout_types}
\title{Types of workouts in strava}
\format{A data table with 8 items and 2 columns
\describe{
  \item{workout_type_id}{Strava Workout type id}
  \item{workout_type}{The workout type name}
}}
\source{
\url{https://developers.strava.com/docs/}
}
\usage{
workout_types
}
\description{
A dataset containing the various type of workout.
}
\keyword{datasets}
