% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect2strava.R
\name{connect2strava}
\alias{connect2strava}
\title{connect2strava}
\usage{
connect2strava(strava.key = NULL, strava.secret = NULL)
}
\arguments{
\item{strava.key}{strava key}

\item{strava.secret}{strave secret}
}
\value{
sig strava token
}
\description{
Create authentication token with strava (based on web authentication)
}
\examples{
\dontrun{
connect2strava(strava.keu = "mykeyhere", strava.secret = "mysecrethere")
}
}
\seealso{
\url{https://www.strava.com/settings/api}
}
