% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-activity.R
\name{getActivity}
\alias{getActivity}
\title{getActivity}
\usage{
getActivity(token = NULL, activity_id = NULL,
  include_all_efforts = FALSE)
}
\arguments{
\item{activity.id}{}
}
\value{
myActivity as list with activity details
}
\description{
Returns the given activity that is owned by the authenticated athlete.
https://developers.strava.com/playground/#/Activities/getActivityById
}
