% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-streams.R
\name{getStreams}
\alias{getStreams}
\title{getStreams}
\usage{
getStreams(activity.id, stream.type = streams.type, path = "./mydata/")
}
\arguments{
\item{path}{}
}
\description{
get all streams from a specific activity and save to a folder
}
