% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listActivities.R
\name{listActivities}
\alias{listActivities}
\title{listActivities}
\source{
\url{https://developers.strava.com/docs/reference/#api-Activities-getLoggedInAthleteActivities}
}
\usage{
listActivities(token = NULL, activities.per.page = 200,
  page.number = 1, after.date = NULL, before.date = NULL)
}
\arguments{
\item{token}{the strava identification token}

\item{activities.per.page}{max number of item to return (must be 200 max)}

\item{after.date}{return activities after date as YYYY-MM-DD}

\item{before.date}{return activities before date as YYYY-MM-DD}
}
\value{
the activities database as json
}
\description{
List all activities from strava api. Returns the activities of an athlete
for a specific identifier. Requires activity:read.
Only Me activities will be filtered out unless requested by a token
with activity:read_all.
}
\examples{
\dontrun{
getActivities(token = myToken, activities.per.page = 200, page.number = 2)
}
}
