% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mountainbikerR.R
\docType{package}
\name{mountainbikeR}
\alias{mountainbikeR}
\alias{mountainbikeR-package}
\title{mountainbiker - when R meets mountainbike and strava}
\description{
A package for mountainbikers who use strava to log their rides and want to
explore and visualise their rides data.
}
\section{mountainbikeR Objective}{

TBD
}

\section{mountainbikeR functions}{


Authentication
\itemize{
  \item \code{\link{connect2strava}} to connect to strave api with key an secret
}

Activities
\itemize{
  \item \code{\link{listActivities}} to get all your activities in one table
  \item \code{\link{cleanActivitiesList}} to clean list of activities
  \item \code{\link{getActivity}} to get all the details of one specific activity
}
Streams
\itemize{
  \item \code{\link{getStreams}} to get streams from an activity
}
Gears
\itemize{
   \item \code{\link{getGear}} to get the list of your gears
}
Toolbox
\itemize{
 \item \code{\link{decodePolyline}} a code to decode polyline (google format for geo path)
 \item \code{\link{extractPath}} to extract path from polyline and build a data.table with id and date
}
}

\section{mountainbikeR datasets}{

The package includes some datasets
\itemize{
 \item \code{\link{stream_types}} contains types of streams
 \item \code{\link{workout_types}} contains types of workout
}
}

