% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6class.R
\name{mountainbiker_initializer}
\alias{mountainbiker_initializer}
\title{R6 Class for a strava application with authorization}
\description{
This strava class is designed to interract with strava API using
a set of methods.
A mountainbikeR object must be initialized using the new method then
all other method will be availabe
}
\details{
TBD
}
\examples{

## ------------------------------------------------
## Method `mountainbiker_initializer$new`
## ------------------------------------------------

\dontrun{
my_mb <- mountainbiker_initializer$new(
    my_app_name = "mycool_app"
    , my_client_id = 77777
    , my_client_secret = "a super secret string"
    , my_app_scope = "read,read_all,activity:read_all" #"public"
    , cache = FALSE
    )
}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{token}}{Shows generated token.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{mountainbiker_initializer$new()}}
\item \href{#method-clone}{\code{mountainbiker_initializer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new moutainbiker object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mountainbiker_initializer$new(
  my_app_name = NULL,
  my_client_id = NULL,
  my_client_secret = NULL,
  my_app_scope = "read,read_all,activity:read_all",
  cache = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{my_app_name}}{chr string, the name of the app defined in strava}

\item{\code{my_client_id}}{chr string, the ID of the app like 27881}

\item{\code{my_client_secret}}{chr string, the secret as defined in strava}

\item{\code{my_app_scope}}{chr string for requested scopes, as a comma delimited}

\item{\code{cache}}{a logical TRUE to keep cached token, false for not.
string, e.g. "activity:read_all,activity:write"}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `mountainbikeR object` object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
my_mb <- mountainbiker_initializer$new(
    my_app_name = "mycool_app"
    , my_client_id = 77777
    , my_client_secret = "a super secret string"
    , my_app_scope = "read,read_all,activity:read_all" #"public"
    , cache = FALSE
    )
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mountainbiker_initializer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
