% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_workout_types.R
\docType{data}
\name{workout_types}
\alias{workout_types}
\title{Availabe workouts in strava API}
\format{
A data frame with 8 rows and 2 variables:
\describe{
  \item{workout_type_id}{the id of the workout type: 11, 10, ...}
  \item{workout_type}{the workout type in plain english : "Ride.Relax",
   "Ride.Race"}
}
}
\source{
\url{https://developers.strava.com/docs/reference/}
}
\usage{
workout_types
}
\description{
Strava activities are associated with a code, describing the type of workout
realised during the activity. This dataset contains the mappin between the
workout types id and their description in plain english.
}
\keyword{datasets}
